requirements.txt
openai==1.79.0
openai-agents
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
logfire
tzlocal
fastapi
uvicorn